# src/lib/api/generated/agent-api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Open Dashboard Agent API
 * API for AI agents to self-register, report status, and manage repos/projects.
 * OpenAPI spec version: 1.0.0
 */
import * as zod from 'zod';

/**
 * @summary List all agents
 */
export const ListAgentsResponseItem = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "identifier": zod.string(),
  "description": zod.string().nullish(),
  "status": zod.enum(['idle', 'working', 'offline']),
  "currentTask": zod.string().nullish(),
  "lastSeenAt": zod.iso.datetime({}).nullish(),
  "createdAt": zod.iso.datetime({})
})
export const ListAgentsResponse = zod.array(ListAgentsResponseItem)


/**
 * @summary Register a new agent
 */




export const CreateAgentBody = zod.object({
  "name": zod.string().min(1),
  "identifier": zod.string().min(1),
  "description": zod.string().optional()
})


/**
 * @summary Get agent details with repos and projects
 */
export const GetAgentParams = zod.object({
  "id": zod.string()
})

export const GetAgentResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "identifier": zod.string(),
  "description": zod.string().nullish(),
  "status": zod.enum(['idle', 'working', 'offline']),
  "currentTask": zod.string().nullish(),
  "lastSeenAt": zod.iso.datetime({}).nullish(),
  "createdAt": zod.iso.datetime({})
}).and(zod.object({
  "repos": zod.array(zod.string()),
  "projectIds": zod.array(zod.string())
}))


/**
 * @summary Update agent fields
 */
export const UpdateAgentParams = zod.object({
  "id": zod.string()
})




export const UpdateAgentBody = zod.object({
  "name": zod.string().min(1).optional(),
  "description": zod.string().optional(),
  "status": zod.enum(['idle', 'working', 'offline']).optional(),
  "currentTask": zod.string().nullish()
})

export const UpdateAgentResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "identifier": zod.string(),
  "description": zod.string().nullish(),
  "status": zod.enum(['idle', 'working', 'offline']),
  "currentTask": zod.string().nullish(),
  "lastSeenAt": zod.iso.datetime({}).nullish(),
  "createdAt": zod.iso.datetime({})
})


/**
 * @summary Update agent status and current task
 */
export const HeartbeatParams = zod.object({
  "id": zod.string()
})

export const HeartbeatBody = zod.object({
  "status": zod.enum(['idle', 'working', 'offline']).optional(),
  "currentTask": zod.string().nullish()
})

export const HeartbeatResponse = zod.object({
  "id": zod.string(),
  "status": zod.string(),
  "currentTask": zod.string().nullish(),
  "lastSeenAt": zod.iso.datetime({})
})


/**
 * @summary List repos tracked by this agent
 */
export const ListAgentReposParams = zod.object({
  "id": zod.string()
})

export const ListAgentReposResponse = zod.object({
  "repos": zod.array(zod.string())
})


/**
 * @summary Replace the list of repos for this agent
 */
export const SetAgentReposParams = zod.object({
  "id": zod.string()
})

export const setAgentReposBodyReposItemRegExp = new RegExp('^[^\/]+\/[^\/]+$');


export const SetAgentReposBody = zod.object({
  "repos": zod.array(zod.string().regex(setAgentReposBodyReposItemRegExp))
})

export const SetAgentReposResponse = zod.object({
  "repos": zod.array(zod.string())
})


/**
 * @summary List projects assigned to this agent
 */
export const ListAgentProjectsParams = zod.object({
  "id": zod.string()
})

export const ListAgentProjectsResponse = zod.object({
  "projectIds": zod.array(zod.string())
})


/**
 * @summary Replace project assignments for this agent
 */
export const SetAgentProjectsParams = zod.object({
  "id": zod.string()
})

export const SetAgentProjectsBody = zod.object({
  "projectIds": zod.array(zod.string())
})

export const SetAgentProjectsResponse = zod.object({
  "projectIds": zod.array(zod.string())
})


/**
 * @summary Get the OpenAPI specification
 */
export const GetOpenApiSpecResponse = zod.looseObject({

})